refactor flashを切り出し
