+++ src-tauri/src/flash.rs
@@ -6,6 +6,8 @@ use std::process::{Command, Stdio};
 use tauri::AppHandle;
 use tauri::Window;
 
+pub mod serial;
+
 #[tauri::command]
 pub fn get_boards_for_select() -> Vec<String> {
     let boards = board_info::get_boards();
@@ -145,25 +147,6 @@ pub async fn flash_to_mcu(
         .stdout(Stdio::piped())
         .spawn()
         .expect("Failed to pio run");
-    let stdout = child.stdout.take().unwrap();
-    // 1行ずつ結果を取る
-    let reader = BufReader::new(stdout);
-    for line in reader.lines() {
-        window.emit("btf-flash-prgoress", line.unwrap()).unwrap();
-    }
-    return Ok(());
-}
-
-#[tauri::command]
-pub fn serial(window: Window) -> Result<(), ()> {
-    /*
-     * pio monitorで見る
-     */
-    let mut child = Command::new("pio")
-        .args(["device", "monitor", "-b", "115200"])
-        .stdout(Stdio::piped())
-        .spawn()
-        .expect("Failed to pio device monitor");
     let stdout = child.stdout.take().unwrap();
     // 1行ずつ結果を取る
     let reader = BufReader::new(stdout);
